separated root query from schema to make code more readable
