+++ server.js
@@ -1,6 +1,6 @@
 const app = require('express')();
 const expressGraphQL = require('express-graphql');
-const schema = require('./schema/schema');
+const schema = require('./schema/rootQuery');
 const _ = require('lodash');
 
 //graphQL middleware
